requirements.txt
pandas>=1.5.0
numpy>=1.21.0
geopy>=2.3.0
openai>=1.0.0
python-dotenv>=0.19.0
reportlab>=3.6.0
matplotlib>=3.5.0
requests>=2.28.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
jinja2>=3.0.0